backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.6,<4.7
PyMySQL>=1.1,<1.2
oss2>=2.19,<2.20
requests>=2.32,<2.33
beautifulsoup4>=4.14,<4.15
openpyxl>=3.1,<3.2
